// src/types/UserProfile.ts
export interface ProfileInterface {
  Amount?: string | number;
  EMAIL?: string;
  ERROR?: string;
  IP?: string;
  Location?: string | null;
  MESSAGE?: string;
  MOBILENUMBER?: string;
  OTPCheck?: string;
  PackageName?: string;
  PopStatus?: string | number;
  ProfileImage?: string;
  RoleId?: string | number;
  RoleName?: string;
  SHOPNAME?: string;
  STATUSCODE?: string;
  SubRoleId?: string | number;
  SubRoleName?: string;
  TOKENID?: string;
  USERID?: string;
  LoginUserId?: string;
  Version?: string | null;
}